+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added option to specify Crosstool-NG version, using `-c` or `--crosstools-toolchain-version`
 - Added an option for [NuShell](https://www.nushell.sh/) completion
+- Add support for LLVM esp-20.1.1_20250829 (#527)
+
 
 ### Fixed